Migrate sync subcommand to the new agent mode
